Quote
The quote was not looking like the desktop version that we had been given so I tried to get it centered but the text wasn't filling out the whole space. I then used text-align justify to stretch it out horizontally but then last center to make the text seem centered.